Removed socket on error handler.  This was probably initialy put in place to deal with issues resulting from issue #2...
Removed socket on error handler.  This was probably initialy put in
place to deal with issues resulting from issue #2 before I know what
they were.  Bumped version as it appears this is now being used in
production by myself and others and I have not intention of changing
the API at this point.